requirements.txt
colorthief==0.2.1
cvzone==1.5.6
Django==4.2.1
djangorestframework==3.14.0
drf_extra_fields==3.4.1
matplotlib==3.7.1
numpy==1.23.5
Pillow==9.5.0
tensorflow==2.12.0
